Update Smart Contracts based on Solidified Audit
Issues:

[1] Ownable contract does not implement a function to revoke access;

Added a `removeOwnership(address allowedAddress)` method to revoke
ownership access. Includes a new event `RevokeOwnership`

[2] Beware of possible collisions in TokenIOStorage;

Reviewed instances where this occurs. There is not an immediate upgrade
to the contracts to check whether a conflict occurs for an issuer. Doing
so would require a check across all issued assets. Handling this issue
may best be suited for UI at the time of authorizing a new firm.

[3] Use of hard coded constants;

Removed all instances of the hard coded "feeMsg" and added it as a
parameter to the TokenIoFeeContract. Now feeContracts can include a
default fee message for transactions that include additional fees (e.g.
for wire transfers, merchants, etc.)

This feeMsg is also able to be set as part of the `setFeeParams()` method.
It is also included in the getFeeParams() method

[4] SafeMath is outdated;

Replaces instances of `assert()` with `require()` to avoid consuming
entire transaction gas

[5] Consider locking version pragma

Locked all version pragma to 0.4.24;

[6] Use of experimental features;

Removed use of experimental features per recommendation of audit

[7] Incorrect constructor comments

Fixed inaccurate constructor comments for contract names

[8] TODO: Consider including messages in require() statements

This is yet to be added, but will be included in a subsequent commit.